fix: id check on deletion

+++ back-end/src/services/clients.service.ts
@@ -37,6 +37,9 @@ async function updateClient(client: ClientId) {
 }
 
 async function deleteClient(id: string) {
+  const checkId = await clientsRepository.findById(id);
+  if (!checkId) throw notFoundError("client not found!");
+
   const deletedClient = await clientsRepository.deleteClient(id);
   return deletedClient;
 }
